+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.1.0] - [date]
+
+### Added
+
+- Added the ability to automatically build a project if no path is specified.
+- Completed HttpService implementation.
+- Completed RunService implementation.
+- Completed LogService implementation.
+
 ## [1.0.0] - 25th Feburary 2024
 
 ### Added